Auth: Reorganize Supabase functions folder for deployment
