+++ supabase/functions/tests/supabaseClientHeaders.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { supabase } from '../../shared/supabase/client.ts'
+import { supabase } from '../../../shared/supabase/client.ts'
 
 describe('supabase client headers', () => {
   it('includes default authorization headers', () => {
